project03_market_place/requirements.txt
django
djangorestframework
psycopg2-binary
celery
redis
channels
channels-redis
beautifulsoup4
requests
daphne
Pillow
django-cors-headers